% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_names.R
\name{compare_names}
\alias{compare_names}
\title{Compares the proportion of children with a given name between 1996 and 2015}
\usage{
compare_names(given_names)
}
\arguments{
\item{given_names}{A string vector containing the names to plot}
}
\value{
A plot comparing the proportion of babies having the specified name(s) between 1996 and 2015
\describe{
\item{Plot}{A ggplot2 plot.}
}
}
\description{
This function provides a plot of the proportion of children with a given name
between the years 1996 and 2015, grouped by sex and name
}
\examples{
# Plot a singular name directly
compare_names("John")

# Plot multiple names directly
compare_names(c("John", "Jane"))

# Plot multiple names using a vector variable
names <- c("John", "Jane", "Jack", "Sam")
compare_names(names)

}
\author{
Julian Blau \email{julian.blau@postgrad.plymouth.ac.uk}, Dan Buckwell \email{dan.buckwell@postgrad.plymouth.ac.uk}, Clint Cridland \email{clint.cridland@postgrad.plymouth.ac.uk}, Charalampos Petrocheilos \email{charalampos.petrocheilos@postgrad.plymouth.ac.uk}
}
