% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_variation.R
\name{name_variation}
\alias{name_variation}
\title{Returns a plot of comparing the number of unique names per year}
\usage{
name_variation()
}
\value{
A plot of the ukbabyname data:
\describe{
\item{Plot}{A ggplot2 plot.}
}
}
\description{
This function returns a plot visualising the number of unique baby names for each year from
1996 to 2015
}
\examples{
name_variation()

}
\author{
Julian Blau \email{julian.blau@postgrad.plymouth.ac.uk}, Dan Buckwell \email{dan.buckwell@postgrad.plymouth.ac.uk}, Clint Cridland \email{clint.cridland@postgrad.plymouth.ac.uk}, Charalampos Petrocheilos \email{charalampos.petrocheilos@postgrad.plymouth.ac.uk}
}
