% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package10593795104317421059735710480219.R
\docType{package}
\name{package_10593795_10431742_10597357_10480219}
\alias{package_10593795_10431742_10597357_10480219}
\alias{package_10593795_10431742_10597357_10480219-package}
\title{package_10593795_10431742_10597357_10480219: A package to provide statistics on the UK EU referendum and Uk Baby Names data}
\description{
This package provides a function to summarize numerically and graphically the data in the Referendum_Vote_Diversity_Score and ukbabynames data sets
}
\author{
Julian Blau \email{julian.blau@postgrad.plymouth.ac.uk}, Dan Buckwell \email{dan.buckwell@postgrad.plymouth.ac.uk}, Clint Cridland \email{clint.cridland@postgrad.plymouth.ac.uk}, Charalampos Petrocheilos \email{charalampos.petrocheilos@postgrad.plymouth.ac.uk}
}
