% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referendum_diversity.R
\name{referendum_diversity}
\alias{referendum_diversity}
\title{Uses the EU referendum data to plot the leave voters against the diversity score}
\usage{
referendum_diversity()
}
\value{
A plot visualising the percentage of leave votes against the diversity of the area
\describe{
\item{Plot}{A ggplot2 plot.}
}
}
\description{
This function provides a plot of the number of leave voters in each voting district
against it's diversity score
}
\examples{
referendum_diversity()

}
\author{
Julian Blau \email{julian.blau@postgrad.plymouth.ac.uk}, Dan Buckwell \email{dan.buckwell@postgrad.plymouth.ac.uk}, Clint Cridland \email{clint.cridland@postgrad.plymouth.ac.uk}, Charalampos Petrocheilos \email{charalampos.petrocheilos@postgrad.plymouth.ac.uk}
}
